requirements/production.txt
kombu==4.6.5
PyYAML==5.1.2
celery==4.3.0
uuid==1.30
python-dateutil==2.8.0
